moving docs/bcdoc -> bcdoc
