+++ awscli/bcdoc/textwriter.py
