+++ awscli/customizations/commands.py
@@ -11,7 +11,7 @@ from awscli.argprocess import unpack_argument, unpack_cli_arg
 from awscli.arguments import CustomArgument, create_argument_model_from_schema
 from awscli.clidocs import OperationDocumentEventHandler
 from awscli.clidriver import CLICommand
-from awscli.docs.bcdoc import docevents
+from awscli.bcdoc import docevents
 from awscli.help import HelpCommand
 from awscli.schema import SchemaTransformer
