+++ tests/unit/customizations/test_commands.py
@@ -16,7 +16,7 @@ import mock
 from awscli.clidriver import CLIDriver
 from awscli.customizations.commands import BasicHelp, BasicCommand
 from awscli.customizations.commands import BasicDocHandler
-from awscli.docs.bcdoc.restdoc import ReSTDocument
+from awscli.bcdoc.restdoc import ReSTDocument
 from botocore.hooks import HierarchicalEmitter
 from tests.unit.test_clidriver import FakeSession, FakeCommand
